+++ src/ipyvizzu/__init__.py
@@ -18,3 +18,28 @@ from .method import Method, Animate, Feature, Store, EventOn, EventOff, Log
 from .json import RawJavaScript, RawJavaScriptEncoder
 from .template import ChartProperty, DisplayTarget, DisplayTemplate
 from .event import EventHandler
+
+__all__ = [
+    "Chart",
+    "Animation",
+    "PlainAnimation",
+    "InferType",
+    "Data",
+    "Config",
+    "Style",
+    "Snapshot",
+    "AnimationMerger",
+    "Method",
+    "Animate",
+    "Feature",
+    "Store",
+    "EventOn",
+    "EventOff",
+    "Log",
+    "RawJavaScript",
+    "RawJavaScriptEncoder",
+    "ChartProperty",
+    "DisplayTarget",
+    "DisplayTemplate",
+    "EventHandler",
+]